requirements.txt
paho-mqtt==2.1.0
opencv-python==4.10.0.84
numpy==2.1.2